+++ crates/config/src/project/mod.rs
@@ -3,21 +3,19 @@
 pub mod global;
 pub mod task;
 
-use crate::constants;
-use crate::errors::{create_validation_error, map_figment_error_to_validation_errors};
+use crate::errors::{create_validation_error, map_validation_errors_to_figment_errors};
 use crate::types::{FileGroups, ProjectID, TaskID};
 use crate::validators::validate_id;
-use figment::value::{Dict, Map};
 use figment::{
     providers::{Format, Serialized, Yaml},
-    Figment, Metadata, Profile, Provider,
+    Error as FigmentError, Figment,
 };
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::Path;
 use task::TaskConfig;
-use validator::{Validate, ValidationError, ValidationErrors};
+use validator::{Validate, ValidationError};
 
 fn validate_file_groups(map: &FileGroups) -> Result<(), ValidationError> {
     for key in map.keys() {
@@ -37,7 +35,7 @@ fn validate_tasks(map: &HashMap<String, TaskConfig>) -> Result<(), ValidationErr
                 return Err(create_validation_error(
                     "required_command",
                     &format!("tasks.{}.command", name),
-                    String::from("An npm/system command is required."),
+                    String::from("An npm/system command is required"),
                 ));
             }
         }
@@ -51,7 +49,7 @@ fn validate_channel(value: &str) -> Result<(), ValidationError> {
         return Err(create_validation_error(
             "invalid_channel",
             "project.channel",
-            String::from("Must start with a #."),
+            String::from("Must start with a `#`"),
         ));
     }
 
@@ -149,32 +147,18 @@ pub struct ProjectConfig {
     pub schema: String,
 }
 
-impl Provider for ProjectConfig {
-    fn metadata(&self) -> Metadata {
-        Metadata::named("Project config").source(constants::CONFIG_PROJECT_FILENAME)
-    }
-
-    fn data(&self) -> Result<Map<Profile, Dict>, figment::Error> {
-        Serialized::defaults(self).data()
-    }
-
-    fn profile(&self) -> Option<Profile> {
-        Some(Profile::Default)
-    }
-}
-
 impl ProjectConfig {
-    pub fn load(path: &Path) -> Result<ProjectConfig, ValidationErrors> {
-        let config: ProjectConfig = match Figment::from(ProjectConfig::default())
-            .merge(Yaml::file(path))
-            .extract()
-        {
-            Ok(cfg) => cfg,
-            Err(error) => return Err(map_figment_error_to_validation_errors(&error)),
-        };
+    pub fn load(path: &Path) -> Result<ProjectConfig, Vec<FigmentError>> {
+        let profile_name = "project";
+        let figment =
+            Figment::from(Serialized::defaults(ProjectConfig::default()).profile(&profile_name))
+                .merge(Yaml::file(path).profile(&profile_name))
+                .select(&profile_name);
+
+        let config: ProjectConfig = figment.extract().map_err(|e| vec![e])?;
 
         if let Err(errors) = config.validate() {
-            return Err(errors);
+            return Err(map_validation_errors_to_figment_errors(&figment, &errors));
         }
 
         Ok(config)
@@ -184,14 +168,14 @@ impl ProjectConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::errors::tests::handled_jailed_error;
+    use crate::constants;
     use moon_utils::string_vec;
     use std::path::PathBuf;
 
     fn load_jailed_config() -> Result<ProjectConfig, figment::Error> {
         match ProjectConfig::load(&PathBuf::from(constants::CONFIG_PROJECT_FILENAME)) {
             Ok(cfg) => Ok(cfg),
-            Err(errors) => Err(handled_jailed_error(&errors)),
+            Err(errors) => Err(errors.first().unwrap().clone()),
         }
     }
 
@@ -238,7 +222,7 @@ fileGroups:
     mod depends_on {
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>dependsOn</id>: Expected a sequence type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a sequence for key \"project.dependsOn\""
         )]
         fn invalid_type() {
             figment::Jail::expect_with(|jail| {
@@ -254,7 +238,7 @@ fileGroups:
     mod file_groups {
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>fileGroups</id>: Expected a map type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a map for key \"project.fileGroups\""
         )]
         fn invalid_type() {
             figment::Jail::expect_with(|jail| {
@@ -268,7 +252,7 @@ fileGroups:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>fileGroups.sources</id>: Expected a sequence type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a sequence for key \"project.fileGroups.sources\""
         )]
         fn invalid_value_type() {
             figment::Jail::expect_with(|jail| {
@@ -326,7 +310,7 @@ tasks:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>tasks</id>: Expected a map type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a map for key \"project.tasks\""
         )]
         fn invalid_type() {
             figment::Jail::expect_with(|jail| {
@@ -340,7 +324,7 @@ tasks:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>tasks.test</id>: Expected struct TaskConfig type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected struct TaskConfig for key \"project.tasks.test\""
         )]
         fn invalid_value_type() {
             figment::Jail::expect_with(|jail| {
@@ -359,7 +343,7 @@ tasks:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>tasks.test.command</id>: Expected a string type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a string for key \"project.tasks.test.command\""
         )]
         fn invalid_value_field() {
             figment::Jail::expect_with(|jail| {
@@ -381,7 +365,7 @@ tasks:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>tasks.test.command</id>: An npm/system command is required."
+            expected = "An npm/system command is required for key \"project.tasks.test.command\""
         )]
         fn invalid_value_empty_field() {
             figment::Jail::expect_with(|jail| {
@@ -405,7 +389,7 @@ tasks:
     mod project {
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project</id>: Expected struct ProjectMetadataConfig type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected struct ProjectMetadataConfig for key \"project.project\""
         )]
         fn invalid_type() {
             figment::Jail::expect_with(|jail| {
@@ -419,7 +403,7 @@ tasks:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project.name</id>: Expected a string type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a string for key \"project.project.name\""
         )]
         fn invalid_name_type() {
             figment::Jail::expect_with(|jail| {
@@ -442,7 +426,7 @@ project:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project.description</id>: Expected a string type, received bool true."
+            expected = "invalid type: found bool true, expected a string for key \"project.project.description\""
         )]
         fn invalid_description_type() {
             figment::Jail::expect_with(|jail| {
@@ -465,7 +449,7 @@ project:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project.owner</id>: Expected a string type, received map."
+            expected = "invalid type: found map, expected a string for key \"project.project.owner\""
         )]
         fn invalid_owner_type() {
             figment::Jail::expect_with(|jail| {
@@ -488,7 +472,7 @@ project:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project.maintainers</id>: Expected a sequence type, received string \"abc\"."
+            expected = "invalid type: found string \"abc\", expected a sequence for key \"project.project.maintainers\""
         )]
         fn invalid_maintainers_type() {
             figment::Jail::expect_with(|jail| {
@@ -511,7 +495,7 @@ project:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>project.channel</id>: Expected a string type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected a string for key \"project.project.channel\""
         )]
         fn invalid_channel_type() {
             figment::Jail::expect_with(|jail| {
@@ -533,7 +517,7 @@ project:
         }
 
         #[test]
-        #[should_panic(expected = "Invalid field <id>project.channel</id>: Must start with a #.")]
+        #[should_panic(expected = "Must start with a `#` for key \"project.project.channel\"")]
         fn channel_leading_hash() {
             figment::Jail::expect_with(|jail| {
                 jail.create_file(
@@ -558,7 +542,7 @@ project:
     mod workspace {
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>workspace</id>: Expected struct ProjectWorkspaceConfig type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected struct ProjectWorkspaceConfig for key \"project.workspace\""
         )]
         fn invalid_type() {
             figment::Jail::expect_with(|jail| {
@@ -572,7 +556,7 @@ project:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>workspace.inheritedTasks</id>: Expected struct ProjectWorkspaceInheritedTasksConfig type, received unsigned int `123`."
+            expected = "invalid type: found unsigned int `123`, expected struct ProjectWorkspaceInheritedTasksConfig for key \"project.workspace.inheritedTasks\""
         )]
         fn invalid_value_type() {
             figment::Jail::expect_with(|jail| {
@@ -591,7 +575,7 @@ workspace:
 
         #[test]
         #[should_panic(
-            expected = "Invalid field <id>workspace.inheritedTasks.include</id>: Expected a sequence type, received string \"abc\"."
+            expected = "invalid type: found string \"abc\", expected a sequence for key \"project.workspace.inheritedTasks.include\""
         )]
         fn invalid_nested_value_type() {
             figment::Jail::expect_with(|jail| {